Refactor layer switcher into component
